requirements.txt
psycopg2-binary
pandas
numpy
faker
python-dotenv
